+++ src/blas/vector/sum.hpp
@@ -1,0 +1,68 @@
+#pragma once
+
+namespace monolish {
+
+namespace {
+template <typename F1> double Dsum_core(const F1 &x) {
+  Logger &logger = Logger::get_instance();
+  logger.func_in(monolish_func);
+
+  double ans = 0;
+  const double *xd = x.data();
+  size_t size = x.size();
+  const size_t xoffset = x.get_offset();
+
+  if (x.get_device_mem_stat() == true) {
+#if MONOLISH_USE_GPU
+#pragma omp target teams distribute parallel for reduction(+ : ans) map (tofrom: ans)
+    for (size_t i = 0; i < size; i++) {
+      ans += xd[i + xoffset];
+    }
+#else
+    throw std::runtime_error(
+        "error USE_GPU is false, but get_device_mem_stat() == true");
+#endif
+  } else {
+#pragma omp parallel for reduction(+ : ans)
+    for (size_t i = 0; i < size; i++) {
+      ans += xd[i + xoffset];
+    }
+  }
+
+  logger.func_out();
+  return ans;
+}
+
+template <typename F1> float Ssum_core(const F1 &x) {
+  Logger &logger = Logger::get_instance();
+  logger.func_in(monolish_func);
+
+  float ans = 0;
+  const float *xd = x.data();
+  size_t size = x.size();
+  const size_t xoffset = x.get_offset();
+
+  if (x.get_device_mem_stat() == true) {
+#if MONOLISH_USE_GPU
+#pragma omp target teams distribute parallel for reduction(+ : ans) map (tofrom: ans)
+    for (size_t i = 0; i < size; i++) {
+      ans += xd[i + xoffset];
+    }
+#else
+    throw std::runtime_error(
+        "error USE_GPU is false, but get_device_mem_stat() == true");
+#endif
+  } else {
+#pragma omp parallel for reduction(+ : ans)
+    for (size_t i = 0; i < size; i++) {
+      ans += xd[i + xoffset];
+    }
+  }
+
+  logger.func_out();
+  return ans;
+}
+
+} // namespace
+
+} // namespace monolish