+++ src/vml/matrix/arithmetic/add2other.sh
@@ -1,0 +1,4 @@
+for func in sub mul div
+do
+  sed "s/add/${func}/g" matadd.hpp > mat${func}.hpp
+done