+++ src/vml/matrix/math_1_3op/max2min.sh
@@ -1,0 +1,1 @@
+cat ./matmax.hpp | sed -e 's/max/min/g'  > matmin.hpp