+++ src/vml/vector/arithmetic/add2other.sh
@@ -1,0 +1,4 @@
+for func in sub mul div
+do
+  sed "s/add/${func}/g" vecadd.hpp > vec${func}.hpp
+done